+++ example/lib/main.dart
@@ -34,142 +34,18 @@ class MyHomePage extends StatefulWidget {
 class _MyHomePageState extends State<MyHomePage> {
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
-        title: Text(widget.title),
-      ),
-      body: const Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.spaceAround,
-          children: <Widget>[
-            PageNavigator(
-              title: "With Blend Effect",
-              child: WithBlendEffect(),
-            ),
-            PageNavigator(
-              title: "Without Blend Effect",
-              child: WithOutBlendEffect(),
-            ),
-            PageNavigator(
-              title: "Custom Decoration",
-              child: CustomDecorationCursor(),
-            ),
-            PageNavigator(
-              title: "Hover Widget Animation",
-              child: WidgetHoverAnimation(),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class WithBlendEffect extends StatelessWidget {
-  const WithBlendEffect({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return AnimatedCursor(
-        enableBlendMask: true,
-        width: 50,
-        height: 50,
-        child: Scaffold(
-          body: Center(
-            child: Text(
-              "Blended Hover Effect",
-              style: Theme.of(context).textTheme.titleLarge,
-            ),
-          ),
-        ));
-  }
-}
-
-class WithOutBlendEffect extends StatelessWidget {
-  const WithOutBlendEffect({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return AnimatedCursor(
-      width: 50,
-      height: 50,
-      decoration:
-          const BoxDecoration(color: Colors.black, shape: BoxShape.circle),
+    return AnimatedMouseCursor(
       child: Scaffold(
-          body: Center(
-        child: Text(
-          "Blended Hover Effect",
-          style: Theme.of(context).textTheme.titleLarge,
+        appBar: AppBar(
+          backgroundColor: Theme.of(context).colorScheme.inversePrimary,
+          title: Text(widget.title),
         ),
-      )),
-    );
-  }
-}
-
-class CustomDecorationCursor extends StatelessWidget {
-  const CustomDecorationCursor({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return AnimatedCursor(
-      decoration:
-          const BoxDecoration(shape: BoxShape.rectangle, color: Colors.red),
-      child: Scaffold(
-          body: Center(
-        child: Text(
-          "Custom Decoration",
-          style: Theme.of(context).textTheme.titleLarge,
-        ),
-      )),
-    );
-  }
-}
-
-class WidgetHoverAnimation extends StatelessWidget {
-  const WidgetHoverAnimation({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return AnimatedCursor(
-      decoration:
-          const BoxDecoration(color: Colors.black, shape: BoxShape.circle),
-      child: Scaffold(
-          body: Center(
-        child: AnimatedCursorMouseRegion(
-          heightMultiplyer: 5, //for increasing cursor size
+        body: Center(
           child: Text(
-            "Hover Widget Animation",
+            "Animated cursor",
             style: Theme.of(context).textTheme.titleLarge,
           ),
         ),
-      )),
-    );
-  }
-}
-
-class PageNavigator extends StatelessWidget {
-  final String title;
-  final Widget child;
-  const PageNavigator({
-    Key? key,
-    required this.title,
-    required this.child,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return ElevatedButton(
-      onPressed: () {
-        Navigator.push(
-            context, MaterialPageRoute(builder: ((context) => child)));
-      },
-      child: Text(
-        title,
-        style: Theme.of(context)
-            .textTheme
-            .bodyLarge!
-            .copyWith(fontWeight: FontWeight.bold),
       ),
     );
   }