+++ lib/animated_mouse_cursor.dart
@@ -2,4 +2,9 @@ library animated_mouse_cursor;
 
 export 'animated_mouse_cursor.dart';
 export 'src/animated_cursor.dart';
+export 'src/animated_cursor_region.dart';
+export 'src/animated_cursor_state.dart';
 export 'src/blended_mask.dart';
+export 'src/constants.dart';
+export 'src/cursor_blend.dart';
+export 'src/src.dart';