+++ lib/src/constants.dart
@@ -1,0 +1,4 @@
+import 'dart:ui';
+
+const black = Color(0xff000000);
+const white = Color(0xffffffff);