+++ CHANGELOG.md
@@ -1,13 +1,26 @@
 # Changelog
 
-## [0.0.1] - 2024-02-08
+## [1.1.0] - 2024-02-18
 
 ### Added
 
-- Initial pre-release
+- animation speed added
+- animated cursor design changed
+
+## [1.0.0] - 2024-02-18
+
+### Added
+
+- web plugin added
 
 ## [0.0.2] - 2024-02-08
 
 ### Added
 
 - Fixed Web Support
+
+## [0.0.1] - 2024-02-08
+
+### Added
+
+- Initial pre-release